requirements.txt
# Core dependencies for SPOF analysis tool
PyGithub==2.1.1
requests==2.31.0
PyYAML==6.0.1
python-dotenv==1.0.0
